Moved stages to Pipeline.ts and deleted ElevateCore. The backend is on the backend branch right now
